Merge: 기프티콘 기능 추가 및 오류 수정

See merge request s09-fintech-finance-sub2/S09P22B207!81

+++ Back-End/src/main/java/com/conseller/conseller/gifticon/dto/request/GifticonRegisterRequest.java
@@ -11,7 +11,7 @@ public class GifticonRegisterRequest {
     private String gifticonBarcode;
     private String gifticonName;
     private String gifticonEndDate;
-    private long subCategory;
-    private long mainCategory;
+    private int subCategory;
+    private int mainCategory;
 
 }
